Seperated claening from sorting
